This pull request is about improving the speed of non-binarized convolutional layers by making the "im2col" operation more efficient. A lot of files got touched, but the changes are mostly pretty minor.

Other work includes:

- Renamed the node that does the "im2col" op from `ReshapeImageNode` to `ReceptiveFieldMatrixNode`.

- Moved `PortMemoryLayout` into model library.

- Improved unit tests somewhat.

- Added 'output directory' option to `darknet_import.py` script

- Various other code cleanup and consolidation.

+++ tools/utilities/debugCompiler/include/ModelComparison.h
@@ -57,9 +57,9 @@ private:
         std::string referenceNodeLabel;
         std::string compiledNodeId;
         std::string compiledNodeLabel;
-        std::vector<size_t> size;
-        std::vector<size_t> stride;
-        std::vector<size_t> offset;
+        std::vector<int> size;
+        std::vector<int> stride;
+        std::vector<int> offset;
         std::vector<float> referenceData;
         std::vector<float> compiledData;
     };
